backend/requirements.txt
fastapi
uvicorn
google-genai
pillow
python-multipart
firebase-admin